+++ bricks/agentic_app/__brick__/{{project_name.snakeCase()}}/lib/core/privacy/consent_service.dart
@@ -1,3 +1,4 @@
+import 'package:{{project_name.snakeCase()}}/core/observability/observability_service.dart';
 import 'package:{{project_name.snakeCase()}}/core/starter/starter_runtime_profile.dart';
 
 class StarterConsentStatus {
@@ -21,6 +22,12 @@ class StarterConsentService implements ConsentService {
 
   @override
   Future<StarterConsentStatus> currentStatus() async {
+    ObservabilityService.instance.log(
+      'starter.consent.snapshot',
+      fields: <String, Object?>{
+        'consent_enabled': StarterRuntimeProfile.consentEnabled,
+      },
+    );
     return StarterConsentStatus(
       analyticsAllowed: !StarterRuntimeProfile.consentEnabled,
       adsAllowed: false,